requirements.txt
dlt[bigquery]>=1.8.1
pandas>=2.0.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
python-dotenv>=1.0.0